memory-chatbot/requirements.txt
streamlit>=1.40.0
openai>=1.58.0
mem0ai>=0.1.29
python-dotenv>=1.0.0
qdrant-client>=1.12.0
